iniciando projeto

+++ src/main/java/com/sistema/gerenciamento_tarefas/service/TaskService.java
@@ -4,8 +4,8 @@ import com.sistema.gerenciamento_tarefas.model.Task;
 import com.sistema.gerenciamento_tarefas.repository.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import java.time.LocalDateTime;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -24,7 +24,6 @@ public class TaskService {
     }
 
     public Task criarTarefa(Task task) {
-        // Inicializa a data de criação ao criar uma tarefa
         task.setDataCriacao(task.getDataCriacao() != null ? task.getDataCriacao() : LocalDateTime.now());
         return taskRepository.save(task);
     }
@@ -44,10 +43,14 @@ public class TaskService {
 
             return taskRepository.save(tarefaExistente);
         }
-        return null; // ou lançar uma exceção, se preferir
+        return null;
     }
 
-    public void deletarTarefa(Long id) {
-        taskRepository.deleteById(id);
+    public boolean deletarTarefa(Long id) {
+        if (taskRepository.existsById(id)) {
+            taskRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
